Group refunds by date in download of accesspay files
This is to match the behaviour of the way the other files
are generated, and reduce use confusion concerning the previously
implemented dynamically ranged accesspay file.

This also switches the download of previous days' files from
querying by batch to querying by date. This means that we now
download all invalid transactions, regardless of current status
i.e. refunded as we all refund_pending.