Cease updating closing balance on statement download
The recorded closing balance for the day is now created when the
transactions are uploaded. The statement closing balance is still, however,
created from the transactions and the opening balance (previous closing).
This is to allow for empty statement cases where no transactions were
uploaded for that day, but will still be sufficient to avoid out of
order bugs.