Change reconciliation period to be from receipt_date to next working day
Previously it was from the receipt_date back to the previous working day.
This means that e.g. the journal downloaded on Monday, for a receipt date
of Friday, will contain the weekend payments, rather than the weekend payments
being in the journal downloaded on Tuesday for a receipt date of Monday.
This is to make the accounting dates for these payments match up
with the dates that they will likely be credited in the prisons.