Add watchdog to dev requirements.txt
On installing the app deependencies, watchdog had to be manually
installed for the app to work. Given the app works in production and
testing without this, adding this to the dev requirements should make
life slightly easier for the next person who has to set up the app.

watchdog is used to monitor code changes and will enable the
browser-sync tool to refresh the site on coce changes.