Removed legacy GA4 setting (`GOOGLE_ANALYTICS_ID`)
`mtp-common` uses this setting to determine whether to send or not data
to legacy GA.

It uses [`getattr(settings, 'GOOGLE_ANALYTICS_ID', None)
`](https://github.com/ministryofjustice/money-to-prisoners-common/blob/199ccaf3507788ba7ea8080033a91be6a4312821/mtp_common/analytics.py#L21)
so it's safe to remove the attribute as it would simply default to `None`
instead of throwing an exception.

NOTE: `bank-admin` doesn't use any of the `Analytics.js` methods *directly*
to send data to legacy GA. So after this there shouldn't be any use of
the legacy GA methods in this application.